refactor: utils 폴더 생성 및 swr 용 fetcher 추가, 파일 구조 변경 #20
