+++ src/app/backtest/detail/page.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import Initialize from "../Initialize";
+import Initialize from "./Initialize";
 import useSWR from "swr";
-import axios from "axios";
 import dynamic from "next/dynamic";
 import { useSearchParams } from "next/navigation";
+import { fetcher } from "@/app/utils/fetcher";
 
 const Financial = dynamic(() => import("../../component/Chart/Financial"), {
   ssr: false,
@@ -17,7 +17,6 @@ const Pie = dynamic(() => import("../../component/Chart/Pie"), {
 export default function BacktestDetailPage() {
   const searchParams = useSearchParams();
   const portfolio = searchParams.get("portfolio");
-  const fetcher = (url: string) => axios.get(url).then((res) => res.data);
   const { data, error, isLoading } = useSWR(
     "http://localhost:80/v1/stock/price?symbol=000050&start_date=2024-04-01&end_date=2024-07-01",
     fetcher
@@ -27,7 +26,7 @@ export default function BacktestDetailPage() {
   if (isLoading) return <div>Loading...</div>;
 
   return (
-    <div className="flex flex-col items-center justify-center mt-10 gap-y-10">
+    <div>
       {portfolio}
       <Initialize portfolio={portfolio ?? ""} />
       {typeof window !== "undefined" && <Pie />}
