+++ src/app/utils/fetcher.ts
@@ -1,0 +1,3 @@
+import axios from "axios";
+
+export const fetcher = (url: string) => axios.get(url).then((res) => res.data);
